trip_tally/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
python-dotenv==1.0.1
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
requests==2.32.3
waitress==3.0.0

